\name{testAffyBatch}
\docType{data}
\alias{testAffyBatch}
\alias{testAffyBatch}
\title{AffyBatch instance testAffyBatch}
\description{
  This is an artifical data set. It contains a 2 genes x 2 samples examples
  (\code{testAffyBatch}) and is suitable for testing the rd-examples in \code{farms}.
}
%\usage{data(testAffyBatch)}
\format{An \code{\link[affy:AffyBatch-class]{AffyBatch}} of 2 samples.}
%\code{\link{cdfenv.example}}
\seealso{\code{\link[affydata]{Dilution}}}
\keyword{datasets}